Small bug fix for bookmarking

+++ server/index.js
@@ -294,6 +294,8 @@ app.get("/recipes", async (req, res) => {
   }
 });
 
+// POST request to bookmark a recipe
+
 app.post("/recipes/:id", async (req, res) => {
   try {
     const cookie = req.signedCookies.authentication; // retrieve authentication cookie from request
@@ -303,18 +305,12 @@ app.post("/recipes/:id", async (req, res) => {
       const user = await pool.query("SELECT * FROM users WHERE Username = $1", [
         cookie,
       ]);
-      if (bookmarked == "1") {
+      if (bookmarked) {
         await pool.query(
           "INSERT INTO recipe_bookmarks (RecipeID, UserID) VALUES($1, $2) RETURNING *",
           [id, user.rows[0].userid]
         );
         res.status(201).send("Created bookmark!");
-      } else {
-        await pool.query(
-          "DELETE FROM recipe_bookmarks WHERE RecipeID = $1 AND UserID = $2",
-          [id, user.rows[0].userid]
-        ); // Delete the database entry
-        res.status(200).send("Bookmarked deleted!");
       }
     } else {
       res.status(401).send("Please register!");
@@ -325,7 +321,7 @@ app.post("/recipes/:id", async (req, res) => {
   }
 });
 
-// Some specified get request
+// GET request to get specified recipe
 
 app.get("/recipes/:id", async (req, res) => {
   try {
@@ -426,29 +422,48 @@ app.get("/recipes/:id", async (req, res) => {
 
 app.delete("/recipes/:id", async (req, res) => {
   try {
-    const { id } = req.params; // Get the URL parameter
-    const recipe = await pool.query(
-      "SELECT * FROM recipes WHERE RecipeID = $1",
-      [id]
-    ); // Retrieve the recipe from the database
-    const images = await pool.query(
-      "SELECT * FROM recipe_instructions WHERE RecipeID = $1",
-      [id]
-    );
-    if (recipe.rows[0]) {
-      // Recipe exists
-      fs.unlinkSync(String(recipe.rows[0].mainimage)); // Delete the main image stored in the database
+    const cookie = req.signedCookies.authentication; // retrieve authentication cookie from request
+    if (cookie) {
+      const { id } = req.params;
+      const { bookmarked } = req.body;
+      if (bookmarked) {
+        const user = await pool.query(
+          "SELECT * FROM users WHERE Username = $1",
+          [cookie]
+        );
 
-      images.rows.forEach((row) => {
-        fs.unlinkSync(String(row.instruction_image));
-      }); // Delete all instruction images
+        await pool.query(
+          "DELETE FROM recipe_bookmarks WHERE RecipeID = $1 AND UserID = $2",
+          [id, user.rows[0].userid]
+        ); // Delete the database entry
+        res.status(200).send("Bookmarked deleted!");
+      } else {
+        const recipe = await pool.query(
+          "SELECT * FROM recipes WHERE RecipeID = $1",
+          [id]
+        ); // Retrieve the recipe from the database
+        const images = await pool.query(
+          "SELECT * FROM recipe_instructions WHERE RecipeID = $1",
+          [id]
+        );
+        if (recipe.rows[0]) {
+          // Recipe exists
+          fs.unlinkSync(String(recipe.rows[0].mainimage)); // Delete the main image stored in the database
 
-      await pool.query("DELETE FROM recipes WHERE RecipeID = $1", [id]); // Delete the database entry
+          images.rows.forEach((row) => {
+            fs.unlinkSync(String(row.instruction_image));
+          }); // Delete all instruction images
 
-      res.status(200).send("Recipe deleted successfully!");
+          await pool.query("DELETE FROM recipes WHERE RecipeID = $1", [id]); // Delete the database entry
+
+          res.status(200).send("Recipe deleted successfully!");
+        } else {
+          // Recipe does not exist
+          res.status(404).send("This recipe does not exist");
+        }
+      }
     } else {
-      // Recipe does not exist
-      res.status(404).send("This recipe does not exist");
+      res.status(401).send("Please register!");
     }
   } catch (error) {
     console.error(error.message);
@@ -487,9 +502,11 @@ app.get("/user/:username", async (req, res) => {
     );
 
     const savedRecipes = await pool.query(
-      "SELECT * FROM recipes WHERE UserID = $1 ORDER BY RecipeID DESC",
+      "SELECT recipes.* FROM recipes WHERE recipes.recipeid IN (SELECT recipeid FROM recipe_bookmarks WHERE userid = $1) ORDER BY recipes.recipeid DESC",
       [id]
     );
+
+    console.log(savedRecipes.rows);
 
     let recipes = {
       reviewed: tempRecipes.rows,
